+++ contracts/TrueCurrencies/mocks/BurnableTokenWithBoundsMock.sol
@@ -4,8 +4,8 @@ pragma solidity 0.6.10;
 import "../BurnableTokenWithBounds.sol";
 
 contract BurnableTokenWithBoundsMock is BurnableTokenWithBounds {
-    constructor(address initialAccount, uint initialBalance) public {
-        _setBalance(initialAccount,  initialBalance);
+    constructor(address initialAccount, uint256 initialBalance) public {
+        _setBalance(initialAccount, initialBalance);
         totalSupply_ = initialBalance;
     }
 }
